requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-rest-framework-mongoengine==3.3.1
django-rest-swagger==2.2.0
djangorestframework==3.9.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mongoengine==0.16.3
openapi-codec==1.3.2
pymongo==3.7.2
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
